refactor: keep code consistency
